backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.1.7
black==24.8.0
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
databases==0.9.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.114.1
filelock==3.16.0
flake8==7.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.8
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.3
pre-commit==3.8.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
virtualenv==20.26.4
